requirements.txt
torch
torchvision
numpy
pyyaml
tqdm
pillow
underthesea
nltk
gradio
